packages/catscan/word/requirements.txt
python-docx==0.8.11
titlecase==2.4
requests==2.28.1